move recent Emoji to its own component
