+++ scripts/index.js
@@ -5,6 +5,7 @@ import { log } from './utils'
 import Search from './components/Search'
 import { CommunityServers, Servers } from './components/Servers'
 import SearchResult from "./components/SearchResult"
+import RecentEmoji from "./components/RecentEmoji"
 
 const BLOBS_ENDPOINT = window.location.host.endsWith('now.sh')
   ? 'https://blobs-gg-test-data.slc.now.sh/data.json'
@@ -22,23 +23,6 @@ function calculateEmojiCount(data) {
   }
 
   return count
-}
-
-function sortEmojiNewest(data) {
-  let blobs = []
-  for (const servers of Object.values(data)) {
-    for (const server of Object.values(servers)) {
-      let se = [...server.emoji]
-      se.map(blob => {
-        blob.server = server.name
-        blob.invite = server.invite
-      })
-      blobs.push(...server.emoji) // get all blobs in a list
-    }
-  }
-  blobs.sort((a, b) => b.id - a.id); // sort blobs, newest first 
-
-  return blobs;
 }
 
 function updatePageState(data) {
@@ -62,8 +46,7 @@ function mount(data) {
 
   log('mounting Recent Emojis')
   const recentEmojis = document.querySelector('#recently-changed-emoji')
-  const sortedEmoji = sortEmojiNewest(data)
-  ReactDOM.render(sortedEmoji.slice(0, 30).map((blob) => <SearchResult key={blob.id} blob={blob} />), recentEmojis)
+  ReactDOM.render(<RecentEmoji data={data} count={30} />, recentEmojis)
   const recentEmojiContainer = document.querySelector('#recently-changed-container')
   recentEmojiContainer.removeAttribute('hidden')
 }
